Add displayData property to UriCell and use preview in UriOverlayEditor
This change adds the displayData property to the UriCell component. The displayData property contains the formatted URI value to be displayed in the cell, and is used instead of the raw URI value in the link area of the UriOverlayEditor component. This allows the URI value to be displayed in a more user-friendly format, such as a shortened URL or email address.